move profile in features/
